requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
plotly>=5.18.0
scikit-learn>=1.3.0
openai>=1.10.0
python-dotenv>=1.0.0
fpdf2>=2.7.0
numpy>=1.24.0
statsmodels>=0.14.0
